refactor ♻️: Code separation (UI & Util)
- Separates code into UI & Util files for future adaption
